requirements.txt
# use pip install to install the following packages :

djangorestframework 
djangorestframework-simplejwt
django-rest-passwordreset